First steps in restoring bitsets
Now that we have phonological rules (and they're running quite slowly), it's time to try out using BitSets again instead of strings.

However, quite a lot has changed since the last time we used BitSets, so this commit does some of the preliminary work.

(1) We eliminated the more complicated output tries in favor of a much simpler one, OutputTrie.  Rather the resurrect the old complicated code, I just make it so OutputTrie is generic and can accept either strings or Tokens.

(2) Expr.generate() and its helper functions had changed somewhat.  The whole Expr.generate structure had become cumbersome in any case, with many "beast" functions duplicating code, and also didn't really need to be inside Expr, so I factored it out and made it generic.

(3) We can now switch between bitset and string implementations with a global, BITSETS_ENABLED.

(4) The naming of things was getting a bit confusing -- e.g. an added "concrete" for some functions that took strings instead of bitset Tokens, but not all of them (generateRandom was concrete).   So now it's consistent: everything taking strings is StringX, everything taking bitset Tokens is BitsetX.

  concreteDeriv -> stringDeriv()
  deriv() -> bitsetDeriv()
  generateDepthFirstConcrete() -> StringGenerator.generateDepthFirst()
  generateDepthFirst() -> BitsetGenerator.generateDepthFirst()
  generateRandom() -> StringGenerator.generateRandom()
  [not implemented] -> BitsetGenerator.generateRandom()