Further refactoring of tapes
Small changes to remove dead code, unused parameters, and further separate the responsibilities of true tapes and TapeCollections.

The only thing of real note is that I removed the parameters and checking where we added an extraneous tapeName param to almost every tape operation.  This was a legacy of the days where we were frequently making any-tape queries (in which you passed a TapeCollection to the derivs as if it were a tape).  We don't do that anymore, and (almost) all uses of this parameter were redundant, it was always `tape.doX(tape.tapeName, ...)`.