Major refactoring of tapes
Fully implementing bitsets again would require major changes to how vocabs work (because of MatchExpr), which would mean major changes to StringTape, or a new class alongside it.  However, the tape hierarchy and how renaming works have gotten convoluted and hard to reason about, and I was loath to complicate it further.  (Having trouble reasoning about renaming was the reason it took us so long to find the nested-replace bug, which was ultimately a tiny error in how tapes get wrapped in RenameTapes.)

The basic source of this confusion is a conflation of responsibilities between StringTapes, TapeCollections, and RenameTapes -- all of these, together, were responsible for keeping track of what each tape was named at any point.  The original reason for this was because we used to be able to make queries (i.e. derivs) without specifying the tape, just saying "any tape", and we handled that by letting the client pass in either a StringTape (if it's a specific tape) or a TapeCollection (if it's not yet a specific tape).  Then, because either can undergo renaming, they both can be wrapped in RenameTapes.  This functionality is no longer needed (we always only query on a specific tape now), but some aspects of it were still used here and there, like during vocab collection/copying.

In any case, there are really two separate responsibilities here, knowing the answers to:

    (Q1) What is the local name, in the current scope, of the tape I care about?
    (Q2) What is the mapping between local tape names and Tape objects?

This refactoring separates those two concerns.  TapeCollection and RenamedTape (now called TapeNamespace and RenamedTapeNamespace respectively) are solely concerned with Q2.  They are no longer kinds of Tapes, and RenamedTapeNamespace only wraps around TapeNamespaces.

Meanwhile, we no longer pass around Tapes during recursive algorithms as the way to identify the tape object we care about.  That was keeping track of both questions simultaneously, what the current name of the tape we care about (through .tapeName), and what tape it referred to (either itself or its child through renaming).  But the TapeNamespace hierarchy now handles the latter question, so all that's really left is keeping track of the current name of the tape we care about.  That's just a string!

So all the deriv/delta functions now take two params in the place of `tape`, a `tapeName:string` param, and a `tapeNS:TapeNamespace` param.  If the deriv/delta really needs a Tape object reference (to access vocabs usually), it can get it through tapeNS.getTape(tapeName).  This might seem indirect but it's a lot simpler and safer than what we were doing... particularly in MatchExpr that can ask about information from multiple tapes, both of which might have undergone renaming.

Be advised: Since we no longer wrap RenameTapes around Tapes, Tape.name now refers to the *global* name of that tape.  It doesn't know it's local name; only TapeNamespaces know that.