Vocab collection bugfixes
In this commit, I was further refactoring the Tapes/TapeNamespaces to better separate their responsibilities, and in the meantime revealed some bugs in vocab collection (see #5).

(1) TapeNamespaces no longer know about vocabs, bits, stringification, or any of that Tape/Token stuff; all they know about is names. If you want to know tape stuff, ask for the Tape.  (I'm open in the future to re-adding these, to let the TapeNamespace be an Edifice so that implementations don't need to know the details of tapes/tokens/etc. But for now it's helpful to keep them separate.)

(2) Tapes no longer know about names; they no longer even know their own name.

(3) Tokens no longer know about stringification; ask the tape to stringify it.  (It's StringTape that knows about vocabulary.)  With that, Tokens no longer know about Tapes, either; that was the last thing in there that referenced it.

(4) It used to be the client's responsibility to check if TapeNS.get(tapeName) had failed, and then decide whether to die or to ignore the failure.  This was probably because it used to be possible to have structures that were ill-formed so far as tapes/renaming were concerned, and you don't want to crash due to a programmer error, you just emit one error message and then otherwise ignore the fact that you're looking for some non-existent tape.  But now we enforce the invariant that all grammars (by the time they get to vocabs/generation) are well-formed w.r.t. tapes.  So now it's genuinely a fatal error to look for a non-existent tape -- an error in OUR code, not the programmer's.

So now, the responsibility for checking and failing lies with TapeNS, and it throws an exception when it fails.

(5) The above revealed a few places where vocabulary collection was attempting to add vocab to the wrong tapes (tapes that don't yet happen to exist at that point in vocab collection).

(5a) HideGrammar, RenameGrammar, and ReplaceGrammar should have been checking to make sure that the tapeName of the currently-being-collected tape isn't equal to their fromTape.  They weren't, so this meant that something inside Hide(X, t2) would still try to add vocab to t2 if t2 happened to exist for some other reason.  (It wouldn't find it, because it can't reach it from in there, but it would still try, and fail silently.)

(5b) CharSetGrammar wasn't checking to see if the currently-being-collected tape was equal to its own tape, so it was probably adding its grammar to every tape.  We only use these rarely, and only for unit testing, so we just never noticed.