Small bugfix and more tape refactoring
(1) In trying out a new way to do renamed TapeNamespaces, I stumbled on a small bug we introduced recently.  You can name the hidden tape in ReplaceGrammars, but there are two checks:

  (a) If the name doesn't start with ".", append "." to the front
  (b) If the name is empty, generate a unique name.

However, we ended up doing (a) before (b), so we ended up with tape names that were ".", and so in some contexts they were not unique.  This commit switches the order so that (b) comes before (a).

(2) The details of how renaming operates on TapeNamespaces are now hidden to clients (i.e., clients don't have to about RenamedTapeNamespace and how to use them).  Instead, clients now just call `tapeNS.rename(from, to)` and the return value is a namespace representing the changed mapping.

(3) Made it so bits are no longer part of the public interface to Tape.  In the resulting moving-stuff-around to make this true, Tape no longer had any non-trivial method implementations, so made it an Interface rather than an abstract class.