Reversed to/from when renaming TapeNamespaces
We've always had a bit of trouble remembering what exactly "to" and "from" refer to when doing tape renaming, and which way (to->from or from->to) to do any particular adjustment.  The recent disentangling of Tapes has made clearer what the ultimate reason is: that there are two different points-of-view to look at renaming.

(1) One is from the inside-out, which is how renaming structures are built from the programmer's point-of-view.  E.g. we have a grammar G, and it has a tape T1, and around that we wrap a Rename that renames tape T1 to tape T2.

(2) The POV is from the outside-in, which is the various recursive algorithms experience renaming.  E.g., the algorithm starts at the global scope of tape names, where that tape is named T2, and when it recurses into a RenameGrammar/RenameExpr that T2 tape changes its name to T1.

RenameGrammars/RenameExprs consistently take the inside-out POV, but for the purposes of algorithms/tapes we were somewhat inconsistent about which POV we took at which point, and this was leading to confusion.  For example, say you had a tapeNS that defined a tape T1, and wanted one where that same tape was named T2.  Do you ask "Rename tape T1 to T2"?  Nope!  You ask "Rename tape T2 to T1", because tapeNS took the POV of the RenameGrammar, not its own POV!  (But meanwhile not everything did or could do that particular switcheroo, so the code appeared inconsistent what was getting renamed in what direction.)

This commit makes tapeNS (and associated functions) consistently take a outside-in POV, consistent with what's actually happening inside tapeNS.  Now, asking tapeNS to "Rename tape T1 to T2" does what you would reasonably expect.

This means that, in functions like (say) RenameExpr.delta, you have to remember that the operations you're doing to tapes go in the opposite direction from the RenameExpr's POV.  The tapeNS's .fromTape is always the RenameExpr's .toTape, and vice-versa.  This is a little tricky but it's consistent and you can eventually wrap your head around it, whereas the old way always led to confusion.  Just remember that during the recursive algorithms you are always going INTO a Rename, and so to get the correct local names you're always REVERSING the effect of that renaming.