Unifying bitset and string derivs
Recent commits have brought the bitset and string derivs to be close enough that many of them are identical, except for their type annotations, and that stringDeriv called stringDeriv and bitsetDeriv called bitsetDeriv.

Now, for Exprs where this is true, both are handled by deriv().  Where the implementations are still different, there's an inherited Expr.deriv() that forwards the query on to the old stringDeriv/bitsetDeriv implementations.

Meanwhile, the naming of Token/Tape classes was still a little misleading.  (I had said earlier that we were consistent about stringX vs bitsetX but this wasn't quite true.  StringTape still involved bitsets!)  Now:

  AbstractToken -> Token
  Token -> BitsetToken
  StringTape -> BitsetTape

This gives room for StringToken and StringTape if necessary later.