Re-enabling LTR parsing/generation
A while back, we experimented with running the algorithm on words "in reverse" (e.g., if handling `t1("hello")`, starting with "o" rather than "h").  In other projects, going back-to-front had helped performance somewhat, and it seemed to help with Michif as well.

Although we had put this in as a parameter (in GenOptions), you couldn't actually change back to LTR; too much had changed in too many places and I never actually got around to making both directions work.  We just kept on doing RTL and never looked back.

With the recent re-unification of stringDeriv and bitsetDeriv meaning that there were now about half as many methods to keep track of, now was high time to buckle down and get this working again.

Important side note: This is unrelated to whether a language's script appears as LTR or RTL on the screen.  That's presentational; in the string in memory both of those languages look the same -- the first letter is on the "left".  This is about the direction that our parsing algorithm takes through memory, regardless of whether characters appear going leftward or rightward on screen.

Changes:

(1) Although we've threaded GenOptions through many places of the code, it wasn't everywhere it needed to be to get this right.  For now I took it out of GenOptions and made it a global.  (As a result, we're not actually using GenOptions anywhere in deriv/delta anymore... but for now I'm keeping it in, we might later.)

(2) ConcatExpr and RTLConcatExpr have been recombined; it's just ConcatExpr now.  ConcatExpr.deriv now considers child1 or child2 to be "first" for the purposes of the algorithm depending on direction.

(3) RTLRepExpr and MatchExpr (which was only RTL as well) have now been generalized to work in either direction.  RTLRepExpr has been renamed "RepeatExpr".  There's really only one thing that changes: both of them spawn off a new expression in concatenation with their child's derivative, but the new expression needs to come first (that is, on the left if LTR and on the right if RTL).  I handle this with a new smart constructor constructOrderedConcat(firstChild, secondChild), which puts the firstChild on the appropriate side.

(4) The only classes that remains different between the two is LiteralExpr and RTLLiteralExpr.  These could be combined but it would be a bit hard to read/understand, so I left them separate.  The responsibility for choosing which one to construct now lies in constructLiteral.