Non-recursive tape renaming
Highly-nested renaming structures (like we get with ordered phonological rules) were causing two kinds of recursion:

  * First, just the recursion to enter the RenameExprs themselves during recursive algorithms like deriv.

  * Secondly, a recursion through the RenamedTapeNamespaces, to grab the Tape objects.

The first of these is hard to avoid, and would be a lot of work to redo without recursion.

But the second (which is, worst-case scenario, exactly as deep as the first) means that we can only go half as deep on renames before blowing the stack.  (For example, say our grammar is just a literal and a big chain of renames.  Abstractly, if we have, say, 101 stack frames before we blow the stack, how many renames can we chain?  Currently, 50.  It takes 51 recursions to get to that LiteralExpr, and then in that literal expression we need to grab the Tape object, and that takes another 50 recursions.  Any longer a chain and we'd blow the stack.)

This has never happened so far... but ultimately, the size of the stack is the critcal, almost-unavoidable constraint on the size of a Gramble grammar, so it probably pays to be stingy about stack frames.

Anyway, this commit finally succeeds in doing TapeNamespace renaming without recursion.  (I'd tried this previously and got it wrong; just forgot to check for tape name collisions.)  Rather than wrap the TapeNamespace in a renaming object (RenamedTapeNamespace) and then recurse down through the chain of renaming objects to get to the original, TapeNamespace.rename just constructs a new TapeNamespace with the appropriately-renamed names.