Unifying verbosity
There were several places verbosity was defined/set and it meant several different things.  For the generation algorithm, it meant debug messages, and was set with a global in utils.  For the Interpreter, it meant compiler and timing messages (like "Collected vocab: 0.024ms"), and was set as a method in the constructor.  The CLI and Sheets sidebar only used verbose in the latter sense.

Both of these senses are now unified using consts defined in utils(SILENT = 0, VERBOSE_TIME = 1, VERBOSE_DEBUG = 2).  I didn't use a Typescript enum for these because verbosity is part of the client-facing API, it's often called from a JS environment, and TS enums compile to something unexpected.  The client programmer shouldn't have to understand what TS enums are underneath to ask for timing messages.

Clients ask for a level of messages in the Interpreter constructor.  (We considered parameterizing every generate method instead, but we decided there wasn't a compelling use case for asking for different levels of verbosity in different invocations of generate from the same Interpreter.  If you really want a higher level of verbosity, go ahead and make another Interpreter, they're cheap compared to generation.)

Interpreter uses this value when compiling the grammar, and then adds it to the opt:GenOptions object, which then is passed to generate().  Currently only generate() itself happens to use the opt.verbose setting, but opt is threaded through all the deriv/delta methods now so you can also use it there if you have debug messages to output.