Minor refactoring of tapes and transforms
A few small annoyances and boilerplate removed:

(1) We were constantly checking Grammar.tape to make sure it was defined, sometimes just for the purposes of linting.  Most of the time, we throw a fairly generic error of "Trying to do something with tapes before calculating them".  The only time we DIDN'T throw an error was in calculateTapes itself.

So now, .tapes is a readonly getter (that can only be string[], not undefined) and the getter is the one responsible for throwing the exception.

The actual member variable is now ._tapes and protected, and is only checked/referenced/assigned-to in .tapes, .calculateTapes, and constructors.

(2) Because GrammarTransform is genericized (to allow the algorithms to pass around args of various types), there wasn't actually any one class/interface that all the transforms shared, meaning a collection of them doesn't have any coherent type.

However, this genericism is only relevant to the execution of the Visitor pattern.  When they're actually used, it's through the same interface, they all have a method `transform(ns: NsGrammar): NsGrammar`.  This commit makes Transform an actual TS interface.

(3) Now we can store Transforms in a list.  Iterating over that list eliminates a lot of repeated boilerplate in Interpreter.

(4) I also make Transforms responsible for describing themselves, for logging purposes; before that was Interpreter's responsibility to know what each transform did.  All Transform has to know now is the order in which they apply.