Code cleanup
(1) createConcat vs createOrderedConcat was confusing me; now the latter is createPrecede.  A "Precede" is like a concat except the left child will always be treated as first, regardless of LTR/RTL.  (There isn't any sort of "PrecedeExpr" or anything, though, everything is ConcatExprs underneath.  It's just a matter of constructing them.)

(2) We had lots of vars that should have been lets, and some vars/lets that should have been consts.  This cleans that up.