Fixed bug in vocab copying
Vocab copying (in which we add all of the vocab of one tape to another tape, e.g. during "vocab bypass" in ReplaceGrammars) was sensitive to the order that we visited the ReplaceGrammars during copyVocab.  We were only ever copying from tape T1 to tape T2 once, but what about if, later, we copied something new from tape T3 to tape T1.  It wouldn't show up on T2 like it should.

Because of how ReplaceGrammars are typically constructed and the order our copyVocab traversal happened in, this wasn't causing any problems.

But if this traversal happened in a different order, we'd get different results.  Testing out something else, I decided to join grammars & rules together in the opposite order (rule as the left child, grammar as the right), which ended up changing the traversal order, causing tests to fail due to incorrect vocabularies.

So now, we do things slightly differently -- collect a directed graph of [fromTape, toTape] edges, then loop through edges until we go through the whole graph without copying anything.  (There's probably a more efficient way to do this, but at this point that'd be premature optimization; it doesn't seem to take much time at all.)