MatchFrom now its own Grammar/Expr
MatchExpr remains one of our most complex expressions to deriv/delta, which makes it hard to reason about when refactoring/debugging/etc.  The interface to it is pretty general (you just wrap it around some child and say "make sure all these tapes match"), but that's what makes it tricky to implement, there's a lot of looping between tapes in a way that we don't do anywhere else.

In practice we only use it in a limited way, through a convenience constructor MatchFrom.  Instead of defining a child X for which all the matching tapes must be defined and have compatible content, child X only has to contain one of the tapes (say, t1), and the rest (say, t2 and t3) are treated as Rename t1->t2 of X and Rename t1->t3 of X, and those things are put into sequence.

This is a bit roundabout way of expressing this, just to meet the more general way Match asserts matches.  So this commit redoes how matching works by making MatchFrom the primitive operation.

The new MatchFrom only takes two tape arguments, a fromTape and a toTape.  (If you want the t1/t2/t3 situation above, nest them.  This is what the convenience constructors do if you give them multiple toTapes.)

If you query on the fromTape, it queries the child and then outputs toTape:nextChar + nextChild.  (That's fundamentally the same as before, but instead of cycling through tapes, checking the child for each of them, outputting a literal for each of them, we just do it once.)

If you query on the toTape, it renames the toTape to fromTape,  queries the child with that, and outputs fromTape:nextChar + nextChild.  (This renaming is new; before the renaming happened in the convenience functions when building the grammar in the first place.  But it does the same thing as before.)

If you query on neither the fromTape nor toTape, it just forwards that on to the child, then wraps and returns like usual.  Note that this results in a different semantics than before.  Before, because of how renaming and sequencing happened, unrelated tapes like this got doubled.  Like MatchFrom(t1:a+t4:b, t1, t2) would result in {t1:a, t2:a, t4:bb}.  Now it's {t1:a, t2:a, t4:b}.  This never came up in real grammars, but it did in tests; those tests have been updated to expect the new semantics.

Vocab works a bit differently than before (with MatchFroms being something that instigates vocab copy), which breaks a ton of Replace tests.  I'll probably just make it work the old way instead, rather than adjust a ton of hard-to-adjust tests.

Other small changes:

(1) Derivs don't typically result in nulls (they just return without generating), it's entirely possible.  Unlike delta (where nulls are common, and we don't bother putting them back on the stack), we were storing these in the stack, and revisiting them later.  It doesn't matter for the outputs (there's nowhere to go from null, we're dead anyway), and probably didn't incur a significant performance penalty, but it clutters up the VERBOSE_DEBUG logs for no reason.

(2) Some new repeat tests, confirming that we handle repeats of nullable complex expressions properly. E.g., (t1:""+t2:"a")* (We do handle those properly, but we weren't testing for it.)

(3) Restored vocab testing lines to the MatchFrom unit tests.