Added tests for replace rule insertions/deletions without context.
Also found and fixed bug in replace rule insertions without context.
The issue was that the expression normally returned from matchAnythingElse()
turns into an expression that matches nothing when the fromExpr is empty.
In this case, the correct behaviour is for matchAnythingElse() to return an
expression matching ".*". We consider fromExpr to be empty if delta of the
fromExpr (with its context) is Epsilon.

There still remains a bug in the insertion/deletion without context with a
repetition count of 2 or more. This is exemplified by replace tests 36j, 36k,
and 36k-2, which fail, but tests 37j and 37k actually suffer from the same bug
even though they pass. Some outputs are not generated. This problem depends on
the tapePriority, which is to say that among the above listed tests, those that
file with the normal tapePriority pass is the tapePriority is reversed, and those
that pass do fail if the tapePriority is reversed. Also, which outputs are missing
depends on whether generation is LTR or RTL.

Further investigation is needed.