First attempts at entanglement
Having both bitset tokens (which represent multiple possible characters at once) and Matches (in which we assert that two characters on different tapes will be the same) leads to a complication: "entanglement".

Midway through a computation, we may have emitted a bitset on one tape, and later, we may have a matching bitset on another tape, but that bitset is further restricted (e.g. by a Join), or just because it's finally time to output a concrete string to the client.  We therefore need to maintain the connection between these two, in the meantime, so that when the later one is finally settled, the earlier one can reflect this.

This commit doesn't yet work out all the details, but I wanted to merge its ancestor branch (matchFrom) into main, so this is just to keep the work-in-progress code safe.