Genericizing TapeNamespace to Namespace
There was nothing much left about Tapes in TapeNamespace, it had just become a wrapper around Map<Tape> with a convenience renaming function.

I need this map-n-rename functionality for (dis)entanglement, so I removed the remaining references to Tape and now it's just Namespace<T>.

Note that this removes the wall of ignorance between Interpreter/Grammars and Vocab (previously they didn't know about Vocab, because TapeNamespace kept those details hidden).  This isn't really a bad thing, though.  I noticed looking through the Interpreter code that it's like "We've got to keep around a TapeNamespace so we remember the vocab so that...". Nah, we should have Interpreter remember what it needs to remember directly, not rely on a Namespace, that's a weird responsibility for a Namespace to have!