Entanglement-aware output decoding
The old OutputTrie.toDict() wasn't compatible with entanglement; it assumed that any possibility of a bitset was viable whenever it encountered one.  What it really needed to do was to be able to temporarily "fix" an interpretation, continue decoding from .prev, and then try another interpretation, etc.

This is much more straightforward to do recursively (otherwise you basically have to implement pseudo-recursion yourself with a stack).  This can't be a long-term solution, though, since outputs can be very long and we'll easily blow the call stack.  So eventually we'll have to reimplement this pseudo-recursively.

This doesn't yet handle multiple entanglements correctly (e.g. where we've matched on 3 tapes).