More general entanglement
There were two missing parts to the previous implementation of entanglement:

(1) We could match(entangled X, unentangled Y) and end up with them entangled, but not match(X, unentangled Y).  Now and/or work back-and-forth: first, we recursively entangle everything all the entanglements from X, then call a new function rightAnd/rightOr to recursively entangle all the entanglements from Y.

In other words, any operation you do with an entangled function causes you to be entangled with it, in any direction.

(2) When decoding, we weren't actually checking whether the value in the entangleRegistry was a valid value for that token.  But it's possible to have invalid values there: for example, when an EntangledToken is joined with another token (fixing its value), but the later token that entangles with it uses that original token (with all possibilities).  So now we check to make sure that the token and all choices in the entangleRegistry are valid.

We're not done yet, however; there are still problems involving nested replacement rules where entanglement is being lost.