Partially fixing negation+match
With the previous commit, nested replaces were failing because negations were stripping the entanglement info from tokens.  But why, there are no neg(match(x)) configurations in those!?

But it's not just when the match is inside negation.  Since entanglement travels along with the token as it traverses the tree, it can pick up an entanglement on one side of the join, then go through a negation on the other side.  This negation loses the entanglement, even though this entanglement isn't entangled with anything inside the negation (and thus should be valid).

So now, negations upwrap any entanglements coming into them, and then rewrap their output tokens in those entanglements afterward.