State-counting verbosity
Generator now counts the number of states it has visited, and can report them with VERBOSE_STATES.

Also, rather than bitmanip and branch on a verbose flag and then console.log(), we now have convenience functions logDebug, logTime, and logStates that handle this.

Also, when BITSETS_ENABLED==true, there are a lot of tests that fail because Match.bitsetDeriv() isn't implemented (and probably will never be, because it'd be hugely complicated and we no longer need it.)  So we now only run those tests when BITSETS_ENABLED==false.