Exposing tape priority as a grammar
The last commit made an Expr-like tape prioritizer but not a grammar-like one; this one fixes that, and so tape priority is now potentially in the hands of test writers.

(1) The Expr-like tape prioritizer wasn't an Expr itself, because it didn't have the same signature for delta/deriv.  In a way this is correct, it's not an Expr like the others, but splitting Expr like this in the program *in general* would be a lot of headache.  So I made it into an Expr (PriorityExpr) and distinguish the sort of tape-specific delta/deriv from the "any-tape" delta/deriv that the prioritizer accepts (we used to call that openQuery or something similar).  The latter are now openDelta/openDeriv.

delta/deriv are just trivial, forwarding on the query to the next.  So if you happen to use a prioritizer elsewhere, it's just ignored.  Possibly in the future we should revisit this, like it'd be nice if prioritization could also potentially happen at a finer level than globally.

(2) There's now a PriorityGrammar and a Priority() convenience function in grammars.ts

(3) Much like CountGrammar, a PriorityGrammar is automatically added to grammars (unless they're already a PriorityGrammar).  I also shifted this into Interpreter (right after the addition of CountGrammar) rather than do it in Generator.

There's an annoying little interaction between the addition of CountGrammar and PriorityGrammar, where we could end up adding a CountGrammar to an existing PriorityGrammar, and then (because the result isn't a PriorityGrammar anymore) adding a PriorityGrammar!  So when adding a CountGrammar we check if we're already a PriorityGrammar and if so, add a CountGrammar around the child rather than around the root.

For both of these, we should really be doing some introspection into the grammar and adding them only when necessary.  (Like, "Oh, you can be infinite, let me start a count" or "Oh you have unprioritized tapes, let me prioritize those.")

(4) There's nothing left in generate.ts that requires knowing about the bitset/string difference, or even having a Generator object.  We're back to there just being a generate() function.