Smarter addition of Counts to infinite grammars
We used to just add a CountGrammar to everything in case it was infinite.  But this isn't great, it's a place where we're setting a hard limit on some grammars that don't need it.  (Like a single-entry grammar that's 1001 characters long doesn't need infinity protection, and would be ruled out by this.)

Now, instead, we consider the grammar and whether it's possible that it be infinite.  Literals, epsilon, null, etc. are finite.  Repeats with maxReps==Infinity are infinite.  Unions and sequences are infinite if any of their children are, joins and equals are infinite if both of their children are.  Counts with a finite maxChars are finite regardless of the finiteness of their children.  Etc.

Recursive grammars are treated as infinite for these purposes, even if in practice we cap recusion.

This will still mark some grammars as infinite when they're really finite, like Rep(Epsilon, 0, Infinity).  Being a bit more clever could address some of these, but there's no general solution.  It'll always be possible to have two infinite grammars where their join is finite, and there's no way to know that in general without having run those grammars.