Changed rule-nesting for faster rules
I was profiling the interaction between tape priority and rule nesting (i.e., whether you structure the joins of grammars and rules like (((G&R1)&R2)&R3) or (G&(R1&(T2&R3))) or whatever).

The results are that the order in which you join them doesn't matter on its own, and only the tape priority matters.  (Of course, right now tape priority is by default determined by grammar structure, so your choice of join order IS crucial, it's just not crucial independent from its effect on tape priority.  If we determined tape priority some other way, it would probably have no effect.)

Assuming the following tapes:

  G has t1
  R1 is from t1 to t2
  R2 is from t2 to t3
  R3 is from t3 to t4

then the best priority is [t1,t2,t3,t4].  This is almost backwards from what we had been using!  In retrospect this makes sense.  The possibilities are forming a tree, and in the reverse order we only prune that tree at the branches, a zillion times; in the correct order we prune that tree at the root.

We had been using a nesting order of (((G&R1)&R2)&R3), which gave us a tape order of [r3,r2,r1,r4].  In order to get the ideal order, we needed to nest (((R3&R2)&R1)&G) or (G&(R1&(R2&R3))).  I chose the latter, and also switched up the order that Join evaluates its components back to left-child-then-right-child, it doesn't seem to matter and is just easier to remember that way.