Fixed vocab collection bug with replace+embed
We encountered a strange bug when putting a replace on the 3rd+ symbol on a sheet, but not on the 1st or 2nd.  WHYYYY?

For example:

X:, text
  , foo

Y:, embed
  , X

Z:, table:, embed
  ,       , X
  ,
  , replace text:, from, to,
  ,              , f,   g

It turns out to be an interaction between embedding, replace, and vocab collection.  We don't visit the same symbol repeatedly on the same tape, because otherwise we spend a ton of time re-traversing the same tables registering the same vocab.

Consider the hidden tape created by this replace, call it ".R16".  That tape is what "text" gets renamed to inside the replace.  So we're collecting vocab, we've created tape ".R16", and now we're going through the grammar collecting the tape.  First we collect X.  That gets no vocab because there's no tape ".R16" here.  Then we collect Y.  In doing so, we collect X again.  And again, no vocab because there's no tape ".R16" here. (Yeah, technically this is repeatedly, it's the 2nd time and thus redundant, but it's only ever twice.)

But then, we collect Z.  In doing so, we change the name ".R16" to "text", and then collect X.  But then we say "Nah already been to X!" and don't collect anything!

So really, during the ".R16" traversal, it's not that we want to store "I've been to X", we want to store "I've been to X as .R16" and "I've been to X as text".

This commit makes collectVocab remember a `[tapeName,symbolName]` pair rather than just `symbolName`, and thus collects the vocabulary properly, and thus avoids the weird replace bug.