Fixing garden path in replacement rules
Our replacement rules were causing exponential runtime behavior, due to a "garden path" in the search tree.

It came down to how we were saying "does not contain X", effectively `~(.*X.*)`.  This worked but had two things that were making it not actually fail until the end of the input.  (With the result that if we joined "abcdef" with "a->A", it would still go down paths that started in "a" even though these would eventually fail.)

The first was just being more aggressive about reducing ~(.*) to the empty set.

Even if we do this, though, there's a second problem, in that `.*X.*` can decide, on one path, to consider X to be part of the first `.*`, and then continue on looking for another X.  This is redundant, the whole thing will have an X whether or not we find a second.

So what we really need is a way to express `.*?X.*`, that is, to be as sparing as possible about what's in that first `.*`.  I'm not sure if there's a way for us to express `.*?` ...what denotation would it have?  But we could express something like `.*?X` in our semantics, it could mean "Every string that has exactly one X and then no other content."

Rather than code this itself, I made a new grammar/expr Short(X), for "short circuit", that has the simple property that it has no derivatives on tape T if it is nullable on tape T.  (So that, in the end, its denotation has no forms where another form would be a proper prefix of it.)  (ShortExpr may not be tape-priority permutation invariant, which is troublesome, but at the moment we're really only using this on a single tape anyway.  Getting basic replacement rules working is priority #1 right now.)

`Short(.*X)` lets us express the property we want, and so not-contains-X is now `~(Short(.*X).*)`.  This now fails at the earliest opportunity, Join(abcdef, a->A) will fail as soon as we try to output "A" rather than fail at the end, eliminating the garden path that was causing exponential behavior.