Re-fixed renamed vocab collection
Before recent fixes, we were only remembering the symbol names of symbols we had visited during vocab collection, and short-circuiting when we visited a symbol we'd already seen.  This turned out to be too strong a short-circuit and we were missing some vocab that way, because you could visit the same symbol masquerading as different names.

So a few commits ago, we started remembering [tapeName,symbolName] pairs.  That's all well and good but oops, tuples are compared by reference!  So it turns out we were never short circuiting, because two different pairs [X,Y] and [X,Y] were never equal.

So we have to convert that tuple to something that's compared by value, like a string.  This commit does that and returns vocab collection to being speedy rather than slow.  (On Michif, the difference is like 100x faster.)