Parallelizing sequences when possible
When you have a sequence like t1:x+t2:y, it doesn't matter what order you visit them in, and either way, the effects of delta/deriv on the irrelevant child are a foregone conclusion.  This isn't a big deal for two, but for a long sequence of single-tape grammars (like in the Michif grammar)... well, actually, it's still not a big deal, but it's a deal.

So this commit introduces ParallelGrammar/Expr, which are collections of single-tape grammars that don't overlap tape-wise.  ParallelExpr is stored as a dict, so that we can just go to the sole grammar appropriate to the query.

This gives us a slight performance boost, maybe 20%.  So not groundbreaking, and it'll be fine to turn it off when we need more clarity about execution, but it doesn't interact with much so it shouldn't interfere with our reasoning too much.